add inputs to utils.ts and refactor files
