+++ Src/UIComponents/utils/utils.ts
@@ -1,3 +1,5 @@
+import { placeholder } from "@babel/types";
+
 export function createElement(tagName: string, classNames?: string[] | string, innerHTML?: string, dataComponent?: string): HTMLElement {
     const element = document.createElement(tagName);
     if(classNames){
@@ -9,17 +11,35 @@ export function createElement(tagName: string, classNames?: string[] | string, i
     return element;
 }
 
-export function generateTextInput(placeholderValue: string, className: string):HTMLInputElement {
-    const textInput = createElement('input', className) as HTMLInputElement;
-    textInput.setAttribute('type', 'text');
-    textInput.setAttribute('placeholder', placeholderValue);
+export function createInput(className: string, type: string, placeholder?: string): HTMLInputElement {
+    const element = createElement('input', className) as HTMLInputElement;
+    element.type = type;
+    if(placeholder) element.placeholder = placeholder;
+    return element;
+}
+
+export function createTextInput(placeholderValue: string, className: string):HTMLInputElement {
+    const textInput = createInput(className, 'text', placeholderValue) as HTMLInputElement;
     textInput.required = true;
     return textInput;
 }
 
-export function createElementInput(name: string, className: string, type: string, placeholder: string): HTMLInputElement {
-    const element = createElement(name, className) as HTMLInputElement;
-    if(type) element.type = type;
-    if(placeholder) element.placeholder = placeholder;
-    return element;
+export function createNumberInput(
+    placeholderValue: string,
+    className: string,
+  ): HTMLInputElement {
+    const numberInput = createInput(className, 'number', placeholderValue);
+    numberInput.setAttribute('min', '0');
+    numberInput.required = true;
+    return numberInput;
 }
+
+export function createRadioInput(name: string, inputValue: string): HTMLInputElement {
+    const radioInput = document.createElement('input');
+    radioInput.setAttribute('type', 'radio');
+    radioInput.setAttribute('name', name);
+    radioInput.setAttribute('value', inputValue);
+    radioInput.setAttribute('id', inputValue);
+    radioInput.required = true;
+    return radioInput;
+  }
